# Fix textareas only after preserved timing