@k0kubun 2017-07-11
As the original problem is heavily tested for both Rails and non-Rails templates and tests pass, I think this is safe to merge.

@haml/committers Thoughts?