@k0kubun 2017-07-14
Since this patch does not introduce any behavior change, merging this for now. Please fell free to add some feedback later.